+++ hash/src/sha512.rs
@@ -9,7 +9,7 @@ pub struct Sha512Hash([u8; SHA512_LENGTH]);
 impl<'a> From<&'a [u8]> for Sha512Hash {
     fn from(slice: &'a [u8]) -> Self {
         assert_eq!(slice.len(), SHA512_LENGTH, "Tried to create instance with slice of wrong length");
-        let mut a = [0 as u8; SHA512_LENGTH];
+        let mut a = [0_u8; SHA512_LENGTH];
         a.clone_from_slice(&slice[0..SHA512_LENGTH]);
         Sha512Hash(a)
     }
@@ -17,7 +17,7 @@ impl<'a> From<&'a [u8]> for Sha512Hash {
 
 impl ::beserial::Deserialize for Sha512Hash {
     fn deserialize<R: ::beserial::ReadBytesExt>(reader: &mut R) -> Result<Self, ::beserial::SerializingError> {
-        let mut a = [0 as u8; SHA512_LENGTH];
+        let mut a = [0_u8; SHA512_LENGTH];
         reader.read_exact(&mut a[..])?;
         Ok(Sha512Hash(a))
     }
